Reimplement to method number 2, as an example
